checking bio Server side, rename Edit Client
