+++ server/routes/profile.js
@@ -9,13 +9,13 @@ const router = express.Router();
 
 /* GETTING MY PROFILE */
 router.get("/mine", isLoggedin, (req, res, next) => {
-    let userid = req.session.user.id;
+    const userid = req.session.user.id;
     //retrieve data from db
     MongoClient.connect(uri, {
         useUnifiedTopology: true,
         useNewUrlParser: true,
     }).then(client => {
-        const users = client.db('twitclone').collection('users');
+        const users = client.db("twitclone").collection("users");
         const projection = { _id: 0, password: 0, email: 0 }; // <--exclusions
         users.findOne({ _id: userid }, { projection: projection }, (err, result) => {
             if (!result) res.sendStatus(404);
@@ -35,7 +35,7 @@ router.get("/user/:userid", (req, res, next) => {
 
 /* handling UPDATE MY PROFILE  */
 router.put("/mine/edit", isLoggedin, (req, res, next) => {
-    let userid = req.session.user.id;
+    const userid = req.session.user.id;
     const { fullname, username, bio } = req.body;
     let errors = []; // input errors
 
@@ -44,6 +44,7 @@ router.put("/mine/edit", isLoggedin, (req, res, next) => {
         let OK = true;
         let reg = /^[ \p{Han}0-9a-zA-Z_\.\'\-]+$/;
         let userReg = /^[0-9a-zA-Z_\S]+$/gi;
+        let bioReg = /^[^<>]+$/gi;
 
         if (!fullname || !username || !bio) {
             //☹ if any empty, END immediately!
@@ -54,13 +55,21 @@ router.put("/mine/edit", isLoggedin, (req, res, next) => {
             errors.push("Name contains illegal characters, ");
             OK = false;
         }
+        if (!bioReg.test(bio)) {
+            errors.push("Bio contains illegal characters, ");
+            OK = false;
+        }
+        if (bio.length > 100) {
+            errors.push("Bio requires ONLY 100 characters max");
+            OK = false;
+        }
         if (!userReg.test(username)) {
             errors.push("Username contains illegal characters, ");
             OK = false;
         }
         return OK;
     };
-    
+
     const checkInputsResult = checkInputs();
     if (checkInputsResult === false) {
         res.status(422).send({ "message": errors, "success": false });
@@ -73,13 +82,12 @@ router.put("/mine/edit", isLoggedin, (req, res, next) => {
             useNewUrlParser: true,
         }).then(client => {
             const users = client.db("twitclone").collection("users");
-            const newValues = { fullname: fullname, username: username, bio: bio};        
+            const newValues = { fullname: fullname, username: username, bio: bio };
             users.updateOne({ _id: userid }, { $set: newValues }, (err, result) => {
                 if (err) {
                     res.status(400).send({ "message": err.message, "success": false });
-                    console.error(err);
+                    console.error("UPDATErr", err);
                 } else {
-                    //RE-WrITE THE SESSION VARIABLES HERE.
                     res.send({ "message": result.modifiedCount, "success": true });
                     console.log("UPDATED", result.result);
                 }
